chore: Include tests and examples in release builds (#872)
Lets build these for release builds. In release builds there is some
optimization added that might help with running tests and seeing
performance numbers. The asserts are removed in release builds so a new
macro was added to check for return status' of the API calls.

Introduced new macro to validate API return statuses (since assert is disabled in release builds).
Added Meson options to toggle building of tests and examples.
Fixed numerous issues with tests and examples to ensure compatibility with release builds.

Signed-off-by: Adit Ranadive <aranadive@nvidia.com>