POSIX backend: LINUXAIO API introduced (#885)
* plugins/posix: getQueueType: error handling fixed

If URING is not supported, the getQueueType() API now returns the
correct return value (queue_t::UNSUPPORTED).

The patch also cleans up the code by removing an unnecessary ifdef.

Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>

* meson.build: POSIX AIO dependency fixed

The current Linux POSIX AIO implementation is provided in user space by
glibc. libaio, on the other hand, is the Linux kernel's AIO access
library.

This patch fixes the incorrect dependency. It will be restored as a
dependency for the Linux AIO plugin in an upcoming patch.

Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>

* plugins/posix: Linux AIO backend introduced

Linux AIO, although only available on Linux, is known for much better
performance than POSIX AIO.

This patch implements a Linux AIO backend and integrates it into the
NIXL build system so it is only built when the platform supports it.

Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>

* benchmark/nixlbench: POSIX backend – LINUXAIO API

This patch adds support for a POSIX LINUXAIO API using the Linux AIO
plugin.

Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>

* POSIX plugin: make Linux AIO the default

This patch makes the Linux AIO API the default for the POSIX backend.

The AIO API now corresponds to the Linux AIO. The POSIX AIO API can now
be selected explicitly by specifying API type POSIXAIO.

Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>

---------

Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>
Co-authored-by: Adit Ranadive <aranadive@nvidia.com>